% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import data}
\usage{
import_data(input_path)
}
\arguments{
\item{input_path}{A filepath of an excel table to import.}
}
\value{
Rmarkdown file.
}
\description{
Import data
}
